examonline/pyproject.toml
[project]
name = "examonline"
version = "0.1.0"
description = "Online Exam Management System with DevOps"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "django>=5.2,<5.3",
    "pillow>=12.0.0",
    "psycopg[binary]>=3.2",
    "pymongo>=4.10",
    "python-dotenv>=1.0",
    "djangorestframework>=3.15",
    "djangorestframework-simplejwt>=5.3",
    "drf-spectacular>=0.27",
    "django-filter>=24.3",
    "django-cors-headers>=4.6",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.8",
    "pytest>=8.3",
    "pytest-django>=4.9",
    "pytest-cov>=6.0",
    "ipython>=8.30",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 100
target-version = "py314"
select = [
    "E",   # pycodestyle errors
    "F",   # pyflakes
    "I",   # isort
    "N",   # pep8-naming
    "W",   # pycodestyle warnings
    "UP",  # pyupgrade
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "DTZ", # flake8-datetimez
]
ignore = [
    "E501",  # line too long (handled by formatter)
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.ruff.lint.isort]
known-first-party = ["examonline", "apps"]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "config.local"
python_files = ["test_*.py", "*_test.py", "tests.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--strict-markers",
    "--tb=short",
    "--cov=apps",
    "--cov-report=term-missing",
    "--cov-report=html",
]
filterwarnings = [
    "ignore::pytest.PytestCollectionWarning",
]
